feat: add BaseManagerTest
