+++ test/resources/withData.csv
@@ -2,7 +2,7 @@ id,type,name,status,description,[epic/subtasks],duration,start-time
 0,TASK,changed name,IN_PROGRESS,desc,10,2024-03-21T20:24:22.679331500
 1,TASK,task2,NEW,desc,10,2024-03-21T20:25:22.679331500
 2,EPIC,Epic1,NEW,desc,[],0,null,null
-3,EPIC,Epic2,DONE,desc,[4 5],120,2024-03-21T20:26:22.694954700,2024-03-21T20:27:32.694954700
+3,EPIC,Epic2,DONE,desc,[4 5],140,2024-03-21T20:26:22.694954700,2024-03-21T20:27:32.694954700
 4,SUBTASK,subTaskDone,DONE,desc,3,10,2024-03-21T20:26:22.694954700
 5,SUBTASK,subTask2,DONE,desc,3,10,2024-03-21T20:27:22.694954700
 0,2,4
